requirements.txt
requests
python-whois
dnspython
